student-chatbot-main/pyproject.toml
[project]
name = "new-project-today"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "google-generativeai>=0.8.6",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]
